Move color map definitions to viz
