+++ conda_package/mpas_tools/viz/SciVisColorColormaps/3Wbgy5.xml
