+++ conda_package/mpas_tools/viz/SciVisColorColormaps/3wave-yellow-grey-blue.xml
