+++ conda_package/mpas_tools/viz/SciVisColorColormaps/4wave-grey-red-green-mgreen.xml
