+++ conda_package/mpas_tools/viz/SciVisColorColormaps/5wave-yellow-brown-blue.xml
