+++ conda_package/mpas_tools/viz/SciVisColorColormaps/blue-3.xml
