+++ conda_package/mpas_tools/viz/SciVisColorColormaps/blue-6.xml
