+++ conda_package/mpas_tools/viz/SciVisColorColormaps/blue-8.xml
