+++ conda_package/mpas_tools/viz/SciVisColorColormaps/brown-5.xml
