+++ conda_package/mpas_tools/viz/SciVisColorColormaps/brown-8.xml
