+++ conda_package/mpas_tools/viz/SciVisColorColormaps/green-7.xml
