+++ conda_package/mpas_tools/viz/SciVisColorColormaps/orange-green-blue-gray.xml
