+++ conda_package/mpas_tools/viz/SciVisColorColormaps/purple-7.xml
