+++ conda_package/mpas_tools/viz/SciVisColorColormaps/purple-8.xml
