+++ conda_package/mpas_tools/viz/SciVisColorColormaps/yellow-7.xml
